\name{techhatch-package}
\alias{techhatch-package}
\alias{techhatch}
\docType{package}
\title{
Collection of utilities provided by Technology Hatchery (short line)
}
\description{
Technology Hatchery uses various R functions internally for such things as processing data and generating predictive analytics.  This package will be periodically but infrequently updated.
}
\details{
\tabular{ll}{
Package: \tab techhatch\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-08-23\cr
License: \tab GNU GPLv3\cr
}

}
\author{
Alfred Wechselberger <alfred.wechselberger@technologyhatchery.com>
}
\keyword{ package }